+++ CHANGES.md
@@ -8,6 +8,7 @@
  * Feature: Remove pandas dependency
  * Breaking Change: Rewrite all rest endpoints to support sync and async versions of the endpoint.
  * Feature: Add dYdX REST endpoints
+ * Feature: Add support for Binance trading REST API
  * Bugfix: Fix typo by renaming rest_options to order_options
 
 ### 1.9.3 (2021-08-05)